requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gradio==5.1.0
gradio-client==1.4.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.3.3
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
propcache==0.2.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pydub==0.25.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
ruff==0.6.9
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.40.0
statsmodels==0.14.4
tenacity==8.5.0
tiktoken==0.8.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
yarl==1.15.3
